blip

+++ src/core/global.rs
@@ -1,7 +1,7 @@
 use std::io::{BufWriter, Write};
 use std::io;
 
-pub type DocId = usize;
+pub type DocId = u32;
 pub type FieldId = u8;
 
 #[derive(Clone,Debug,PartialEq,PartialOrd,Eq,Hash)]
